Refactor Tab UI to component
