+++ main.js
@@ -69,6 +69,8 @@ function createWindow(commandLineArguments, workingDirectory) {
         })
     })
 
+    mainWindow.webContents.openDevTools()
+
     ipcMain.on('rebuild-menu', function(_evt, loadInit) {
         // ipcMain is a singleton so if there are multiple Oni instances
         // we may receive an event from a different instance
